third_party/WebKit/LayoutTests/fast/dom/gc-stylesheet-rule-property.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<head>
<style>
.doesntexit { display: none }
</style>
<script>

test(() => {
  document.styleSheets[0].cssRules[0].custom_property = 'persist';
  gc();
  assert_equals(
      "persist",
      document.styleSheets[0].cssRules[0].custom_property,
      "Should not loose property on GC.")
}, 'Tests properties on style sheet rules are not lost. See http://crbug.com/830910.');

</script>
</head>
<body>
</body>
</html>
